Move getPncQueryTimeFromAho to tests/helpers/comparison folder
