+++ packages/core/comparison/lib/comparePhase1.ts
@@ -14,8 +14,8 @@ import serialiseToXml from "../../lib/serialise/ahoXml/serialiseToXml"
 import getMessageType from "../../phase1/lib/getMessageType"
 import phase1Handler from "../../phase1/phase1"
 import generateMockPncQueryResultFromAho from "../../tests/helpers/comparison/generateMockPncQueryResultFromAho"
+import getPncQueryTimeFromAho from "../../tests/helpers/comparison/getPncQueryTimeFromAho"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
-import getPncQueryTimeFromAho from "./getPncQueryTimeFromAho"
 import isIntentionalDifference from "./isIntentionalDifference"
 import parseIncomingMessage from "./parseIncomingMessage"
 import { sortExceptions } from "./sortExceptions"
