+++ packages/core/comparison/lib/comparePncMatching.ts
@@ -9,8 +9,8 @@ import CoreAuditLogger from "../../lib/auditLog/CoreAuditLogger"
 import { parseAhoXml } from "../../lib/parse/parseAhoXml"
 import CorePhase1 from "../../phase1/phase1"
 import generateMockPncQueryResultFromAho from "../../tests/helpers/comparison/generateMockPncQueryResultFromAho"
+import getPncQueryTimeFromAho from "../../tests/helpers/comparison/getPncQueryTimeFromAho"
 import MockPncGateway from "../../tests/helpers/MockPncGateway"
-import getPncQueryTimeFromAho from "./getPncQueryTimeFromAho"
 import parseIncomingMessage from "./parseIncomingMessage"
 import summariseMatching from "./summariseMatching"
