+++ packages/core/tests/helpers/comparison/getPncQueryTimeFromAho.ts
@@ -1,6 +1,6 @@
 import { XMLParser } from "fast-xml-parser"
 
-import type { AhoXml } from "../../types/AhoXml"
+import type { AhoXml } from "../../../types/AhoXml"
 
 const getPncQueryTimeFromAho = (ahoXml: string): Date | undefined => {
   const parser = new XMLParser({
