+++ packages/core/tsconfig.build.json
@@ -10,6 +10,7 @@
     "phase3/tests/helpers/getPncOperationsFromPncUpdateDataset.ts",
     "tests/helpers/MockPncGateway.ts",
     "tests/helpers/comparison/generateMockPncQueryResultFromAho.ts",
+    "tests/helpers/comparison/getPncQueryTimeFromAho.ts",
     "tests/types/ComparisonFile.ts",
   ]
 }
